ASP-2810 rearrange auth module on jobbergate-core
