+++ jobbergate-core/jobbergate_core/__init__.py
@@ -1,7 +1,13 @@
 """
 Jobbergate-core contains key components that are shared among sub-projects.
 """
+from .auth import AuthenticationError, JobbergateAuth, TokenError
 from .version import __version__
 
 
-__all__ = ["__version__"]
+__all__ = [
+    "__version__",
+    "AuthenticationError",
+    "JobbergateAuth",
+    "TokenError",
+]
