+++ jobbergate-core/jobbergate_core/auth/core.py
@@ -1,160 +1,46 @@
 """
 Utilities for handling auth in Jobbergate.
 """
-from __future__ import annotations
-
 import time
-from dataclasses import dataclass, field, replace
-from enum import Enum
+from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Any, Dict, List
+from typing import Dict
 
-import buzz
 import httpx
-import pendulum
-from jose.jwt import decode
 from loguru import logger
 
-
-class AuthenticationError(buzz.Buzz):
-    """
-    Exception for errors related to authentication.
-    """
-
-    pass
-
-
-class TokenError(AuthenticationError):
-    """
-    Exception for errors related to tokens.
-    """
-
-    pass
-
-
-class TokenType(str, Enum):
-    """
-    Types of tokens.
-    """
-
-    ACCESS = "access"
-    REFRESH = "refresh"
-
-
-@dataclass(frozen=True)
-class Token:
-    """
-    Class for handling tokens.
-    """
-
-    content: str
-    cache_directory: Path
-    label: TokenType
-    file_path: Path = field(init=False, hash=False, repr=False)
-    data: Dict[str, Any] = field(
-        default_factory=dict,
-        init=False,
-        hash=False,
-        repr=False,
-    )
-
-    def __post_init__(self):
-        """
-        Post init method.
-        """
-        TokenError.require_condition(bool(self.content), "Token content is empty")
-
-        # There is one point to keep in mind when using frozen=True
-        # see https://docs.python.org/3/library/dataclasses.html#frozen-instances
-        object.__setattr__(
-            self,
-            "file_path",
-            self.cache_directory / f"{self.label}.token",
-        )
-        data = self._get_metadata()
-        object.__setattr__(self, "data", data)
-
-    def _get_metadata(self) -> Dict[str, Any]:
-        """
-        Extract the data from the token.
-        """
-        logger.debug(f"Getting data from {self.label} token")
-
-        with TokenError.handle_errors("Unable to extract data from the token"):
-            data = decode(
-                token=self.content,
-                key="",
-                options=dict(
-                    verify_signature=False,
-                    verify_aud=False,
-                    verify_exp=False,
-                ),
-            )
-
-        return data
-
-    @classmethod
-    def load_from_cache(cls, cache_directory: Path, label: TokenType) -> Token:
-        """
-        Alternative initialization method that loads the token from the cache.
-        """
-        file_path = cache_directory / f"{label}.token"
-        logger.debug(f"Loading token from {file_path.as_posix()}")
-
-        TokenError.require_condition(file_path.exists(), "Token file was not found")
-
-        with TokenError.handle_errors("Unknown error while loading the token"):
-            content = file_path.read_text().strip()
-
-        return cls(content=content, cache_directory=cache_directory, label=label)
-
-    def save_to_cache(self):
-        """
-        Save the token to the cache.
-        """
-        logger.debug(f"Saving token to {self.file_path}")
-        TokenError.require_condition(
-            self.file_path.parent.exists(),
-            "Parent directory does not exist",
-        )
-
-        with TokenError.handle_errors("Unknown error while saving the token"):
-            self.file_path.write_text(self.content.strip())
-            self.file_path.chmod(0o600)
-
-    def clear_cache(self):
-        """
-        Clear the cache.
-        """
-        logger.debug(f"Clearing cached token from {self.file_path}")
-        if self.file_path.exists():
-            self.file_path.unlink()
-
-    def is_expired(self) -> bool:
-        """
-        Check if the token is expired.
-        """
-        logger.debug(f"Checking if {self.label} token has expired")
-        token_expiration = self.data.get("exp", 0)
-        TokenError.require_condition(token_expiration > 0, "The expiration date was not found")
-
-        current_time_UTC = pendulum.now().int_timestamp
-        is_expired = token_expiration >= current_time_UTC
-        logger.debug(f"    Token is expired: {is_expired}")
-
-        return is_expired
-
-    def replace(self, **changes) -> Token:
-        """
-        Create a new instance of the token with the changes applied.
-        """
-        return replace(self, **changes)
+from .exceptions import AuthenticationError, TokenError
+from .token import Token, TokenType
 
 
 @dataclass
 class JobbergateAuth:
     """
     Class for handling authentication in Jobbergate.
+
+    Arguments:
+        cache_directory (Path): Directory to be used for the caching tokens.
+        login_domain (str): Domain used for the login.
+        login_audience (str): Audience of the login.
+        login_client_id (str): Client ID used for login.
+        tokens (Dict[TokenType, Token]): Dictionary holding the tokens needed for authentication.
+
+    Examples:
+        >>> from pathlib import Path
+        >>> import requests
+        >>> from jobbergate_core import JobbergateAuth
+        >>> jobbergate_auth = JobbergateAuth(
+        ...     cache_directory=Path("."),
+        ...     login_domain="http://keycloak.local:8080/realms/jobbergate-local",
+        ...     login_audience="https://local.omnivector.solutions",
+        ...     login_client_id="cli",
+        ... )
+        >>> jobbergate_base_url = "http://localhost:8000/jobbergate"
+        >>> jobbergate_auth.acquire_tokens()
+        Login Here: http://keycloak.local:8080/realms/jobbergate-local/device?user_code=LMVJ-XOLG
+        >>> response = requests.get(f"{jobbergate_base_url}/applications", auth=jobbergate_auth)
+        >>> response.raise_for_status()
+        >>> print(f"response = {response.json()}")
     """
 
     cache_directory: Path
@@ -171,7 +57,7 @@ class JobbergateAuth:
 
         self.acquire_tokens()
 
-        access_token = self.tokens.get(TokenType.ACCESS)
+        access_token = self.tokens.get(TokenType.ACCESS, "")
         AuthenticationError.require_condition(access_token, "Access token was not found")
 
         request.headers["Authorization"] = f"Bearer {access_token.content}"
@@ -182,7 +68,7 @@ class JobbergateAuth:
         Acquire the tokens.
         """
         logger.debug("Acquiring tokens")
-        self.load_from_cache(force=False)
+        self.load_from_cache(skip_loaded=True)
         if TokenType.ACCESS in self.tokens and not self.tokens[TokenType.ACCESS].is_expired():
             return
         elif TokenType.REFRESH in self.tokens and not self.tokens[TokenType.REFRESH].is_expired():
@@ -190,14 +76,14 @@ class JobbergateAuth:
             return
         self.login()
 
-    def load_from_cache(self, force: bool = False):
+    def load_from_cache(self, skip_loaded: bool = True):
         """
         Load the tokens from the cache.
         """
-        logger.debug(f"Loading tokens from cache directory: {self.cache_directory.as_posix()}")
+        logger.debug("Loading tokens from cache directory: {}", self.cache_directory.as_posix())
 
         for t in TokenType:
-            if t in self.tokens and not force:
+            if t in self.tokens and skip_loaded:
                 continue
             try:
                 new_token = Token.load_from_cache(self.cache_directory, label=t)
@@ -210,6 +96,7 @@ class JobbergateAuth:
         """
         Save the tokens to the cache.
         """
+        self.cache_directory.mkdir(parents=True, exist_ok=True)
         for token in self.tokens.values():
             token.save_to_cache()
 
@@ -242,8 +129,11 @@ class JobbergateAuth:
                     client_id=self.login_client_id,
                 ),
             )
-            if response.status_code == "ok":
+            try:
+                response.raise_for_status()
                 break
+            except httpx.HTTPStatusError:
+                continue
 
         self._process_tokens_from_response(response)
